pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scpepato"
version = "0.1.0"
description = "Single-Cell Perturbation Analysis Toolkit for Optical Pooled Screening"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Marco DiBernardo", email = "mdiberna@mit.edu"}
]

dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "pyarrow",
    "anndata",
    "scanpy",
    "numba>=0.57",
    "umap-learn",
    "phate",
    "leidenalg",
    "matplotlib",
    "seaborn",
    "scikit-learn",
    "tqdm",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "ipykernel",
    "jupyterlab",
]
torch = [
    "torch",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
